+++ src/main/java/http/Response.java
@@ -1,4 +1,4 @@
-package webserver.support;
+package http;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
