+++ src/test/java/http/RequestTest.java
@@ -1,4 +1,4 @@
-package webserver.support;
+package http;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
